Remove DLMALLOC_DEBUG and abort() override
dlmalloc already has `DEBUG` macros that is used to enable debugging, so
just use that one for consistency.

In terms replacing abort with direct call to __builtin_unreachable this
seems a little odd to do that here but not in other places.

`abort` is a call to an import so is also a single instruction, and the
JS version of abort has almost no chance of being completely avoided
(even the smallest program will likely reference it).  Perhaps this
comment dates back to a different time when abort was implemented
differently or perhaps it refers to the difference is size between a
`call` instruction and a `trap` instruction in the binary.  However, the
ABORT macro is mostly only used in the `DEBUG` version of dlmalloc so
this argument doesn't make much sense.

See #15662 and #15628